refactor: use prefixed classnames instead of inline styles, render <style>
this commit changes how we use styles from the default react
inline styles to CSS. There are a couple drawbacks to using inline
styles, one is that we can’t use nested selectors (like *) as well as
hover, focus etc. states as well as pseudo elements.

We render our own <style> tag with the component
and all our classnames are prefixed to prevent conflicts with the
surrounding site.