don’t bubble events
bubbling was necessary in an earlier implementation where the event
was dispatched on the React component rather than the custom element
itself.

The event needed to be `composed` and bubble for it to cross over the
Shadow DOM boundary into the hosting DOM, but we are dispatching the
event directly on the custom element so we don’t need that anymore.

This also allows us to remove the event prefix as no bubbling means that
it can’t conflict with other events that are listened to higher up the
chain.