build: switch to ESM, add external dep build
this commit updates the output format (yet again) to use ESM. It also
adds a second output file to be used when installing and importing
the component via npm. This build has all dependencies marked as
external as it is expected to be bundled again.

The new external dependency build is set as the `module` entrypoint in
the package.json. The everything included bundle, `index.wcb.js`, is
expected to be used via a script tag directly in the browser.

WCAutocomplete is exported as default now too, which is ignored for the
wcb build but can be useful for the importable build.