refactor: css: replace css module with custom esbuild plugin
This commit replaces the esbuild-css-modules-plugin witb a much smaller
and simpler solution: the plugin uses esbuild itself to bundle the CSS
and then returns the CSS string using the built-in `text` loader, which
means it gets inlined during the build.

Aside from less dependencies and less complexity, this also has the
advantage that the CSS is now minified, which reduces the bundle size
(granted not by a lot as we don’t have a lot of CSS but still).