customization: add string template
this template allows customization of the way a Feature is converted
into a string and can be used for "smaller" customization where a fully
new template with custom styles would be overkill.

Also this applies to the resulting string that is placed in the input
field after an item is selected.

this commit also refactors the way templates are passed to parse them
both in a single function instead of duplicating the logic. This also
allows easily adding more templates in the future.