don’t trim search term
trimming the search term makes `uk` and `uk ` the same query, while in
fact the space at the end indicates that it is a standalone word (for
example because the user wants to search in the united kingdom and not
continue typing `ukraine`). This way the server can handle that
correctly.